line-bot/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
line-bot-sdk==3.14.0
stripe==10.5.0
gunicorn==22.0.0
